Update MicrosoftTeams.munki.recipe
Overrides the package version when calling the MunkiImporter processor. The makepkginfo code incorrectly grabs the version of the com.microsoft.MSTeamsAudioDevice component package instead of the version of the com.microsoft.teams component package when detecting the overall pkg version. Since we have the "correct" version available to use already, just tell makepkginfo to use it.