Update Fiji.download.recipe
Current recipe is finding the latest ImageJ version but Fiji has a different versioning scheme (and often is a couple of ImageJ versions behind so even the version of ImageJ bundled is not what is being pulled).  This changes where Fiji version is sourced from to the Maven repository metadata.  It does change it significantly from the ImageJ version (being 1.54j or something) to the Fiji version (2.15.0 or something).